safari-534.53-branch/LayoutTests/fast/dom/Window/window-custom-prototype-expected.txt

Test what happens when you set the window's prototype to various values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS __proto__ = window; __proto threw exception Error: cyclic __proto__ value.
PASS __proto__ = chainPointingBackToWindow; __proto__ threw exception Error: cyclic __proto__ value.
PASS __proto__ = 1; __proto__ is originalWindowPrototype
PASS __proto__ = 'a string'; __proto__ is originalWindowPrototype
PASS __proto__ = anotherObject; __proto__ is anotherObject
PASS anotherObject.__proto__ = window; __proto__ threw exception Error: cyclic __proto__ value.
PASS __proto__ = null; __proto__ is null
PASS __proto__ = 1; __proto__ is null
PASS __proto__ = 'a string'; __proto__ is null
PASS __proto__ = anotherObject; __proto__ is anotherObject
PASS __proto__ = originalWindowPrototype; __proto__ is originalWindowPrototype
PASS anotherObject.__proto__ = window; anotherObject.__proto__ is window
PASS successfullyParsed is true

TEST COMPLETE
